+++ src/parseDdt.h
@@ -66,11 +66,10 @@ static inline int add_deState(nodeInfo ni, char *name, char *v, int hasLhs, int 
     trans_syntax_error_report_fn(NODOT);
   }
   new_or_ith(v);
-  if (!rx_syntax_allow_assign_state &&
-      ((tb.ini[tb.ix] == 1 && tb.ini0[tb.ix] == 0) ||
+  if (((tb.ini[tb.ix] == 1 && tb.ini0[tb.ix] == 0) ||
        (tb.lh[tb.ix] == isLHS || tb.lh[tb.ix] == isLHSparam))){
     updateSyntaxCol();
-    sPrint(&_gbuf,_("cannot assign state variable %s; For initial condition assignment use '%s(0) = #'.\n  Changing states can break sensitivity analysis (for nlmixr glmm/focei).\n  To override this behavior set 'options(rxode2.syntax.assign.state = TRUE)'"),v,v);
+    sPrint(&_gbuf,_("cannot assign state variable %s; For initial condition assignment use '%s(0) = #'."),v,v);
     trans_syntax_error_report_fn0(_gbuf.s);
   }
   tb.lh[tb.ix] = isState;